requirements.txt
streamlit>=1.24.0
ultralytics>=8.0.0
Pillow>=9.0.0
opencv-python>=4.7.0
numpy>=1.22.0
